fix: cleanup part 3
