+++ packages/shared/src/features/shortcuts/components/WebappShortcutsRow.tsx
@@ -11,27 +11,22 @@ import {
 } from '@dnd-kit/core';
 import type { DragEndEvent } from '@dnd-kit/core';
 import {
-  arrayMove,
   horizontalListSortingStrategy,
   SortableContext,
   sortableKeyboardCoordinates,
 } from '@dnd-kit/sortable';
 import { useSettingsContext } from '../../../contexts/SettingsContext';
 import { useLogContext } from '../../../contexts/LogContext';
-import { useLazyModal } from '../../../hooks/useLazyModal';
-import { useToastNotification } from '../../../hooks/useToastNotification';
-import { LazyModal } from '../../../components/modals/common/types';
 import { LogEvent, ShortcutsSourceType, TargetType } from '../../../lib/log';
 import { ShortcutTile } from './ShortcutTile';
 import { AddShortcutTile } from './AddShortcutTile';
-import { useShortcutsManager } from '../hooks/useShortcutsManager';
 import {
   useDragClickGuard,
   DRAG_ACTIVATION_DISTANCE_PX,
 } from '../hooks/useDragClickGuard';
-import { useShortcutDropZone } from '../hooks/useShortcutDropZone';
 import { DEFAULT_SHORTCUTS_APPEARANCE } from '../types';
-import type { Shortcut, ShortcutsAppearance } from '../types';
+import type { ShortcutsAppearance } from '../types';
+import { useManualShortcutsRow } from '../hooks/useManualShortcutsRow';
 
 interface WebappShortcutsRowProps {
   className?: string;
@@ -45,15 +40,13 @@ export function WebappShortcutsRow({
   className,
 }: WebappShortcutsRowProps): ReactElement | null {
   const { flags, showTopSites } = useSettingsContext();
-  const { openModal } = useLazyModal();
-  const { displayToast } = useToastNotification();
   const { logEvent } = useLogContext();
-  const manager = useShortcutsManager();
+  const manualRow = useManualShortcutsRow();
 
   const enabled = flags?.showShortcutsOnWebapp ?? false;
   const appearance: ShortcutsAppearance =
     flags?.shortcutsAppearance ?? DEFAULT_SHORTCUTS_APPEARANCE;
-  const shortcuts = manager.shortcuts;
+  const shortcuts = manualRow.shortcuts;
 
   const loggedRef = useRef(false);
   useEffect(() => {
@@ -98,43 +91,13 @@ export function WebappShortcutsRow({
     if (!over || active.id === over.id) {
       return;
     }
-    const urls = shortcuts.map((s) => s.url);
-    const oldIndex = urls.indexOf(active.id as string);
-    const newIndex = urls.indexOf(over.id as string);
-    if (oldIndex < 0 || newIndex < 0) {
-      return;
-    }
-    manager.reorder(arrayMove(urls, oldIndex, newIndex));
+    manualRow.reorderShortcuts(active.id as string, over.id as string);
   };
-
-  const onEdit = (shortcut: Shortcut) =>
-    openModal({
-      type: LazyModal.ShortcutEdit,
-      props: { mode: 'edit', shortcut },
-    });
-
-  const onRemove = (shortcut: Shortcut) => manager.removeShortcut(shortcut.url);
-
-  const onAdd = () =>
-    openModal({ type: LazyModal.ShortcutEdit, props: { mode: 'add' } });
-
-  const onDropUrl = async (url: string) => {
-    const result = await manager.addShortcut({ url });
-    if (result.error) {
-      displayToast(result.error);
-    }
-  };
-
-  const canAcceptDroppedUrl = manager.canAdd;
-  const { isDropTarget, dropHandlers } = useShortcutDropZone(
-    onDropUrl,
-    canAcceptDroppedUrl,
-  );
 
   if (!enabled || !showTopSites) {
     return null;
   }
-  if (shortcuts.length === 0 && !manager.canAdd) {
+  if (shortcuts.length === 0 && !manualRow.canAdd) {
     return null;
   }
 
@@ -145,14 +108,14 @@ export function WebappShortcutsRow({
       onClickCapture={suppressClickCapture}
       onAuxClickCapture={suppressClickCapture}
       onDragStartCapture={suppressNativeDragCapture}
-      {...dropHandlers}
+      {...manualRow.dropHandlers}
       className={classNames(
         'hidden flex-wrap items-center mobileXL:flex',
         appearance === 'tile' && 'items-start gap-x-1 gap-y-2',
         appearance === 'icon' && 'gap-1',
         appearance === 'chip' && 'gap-1',
         'rounded-12 transition-[box-shadow,background-color] duration-150 motion-reduce:transition-none',
-        isDropTarget &&
+        manualRow.isDropTarget &&
           'bg-overlay-float-cabbage ring-2 ring-accent-cabbage-default ring-offset-4 ring-offset-background-default',
         className,
       )}
@@ -173,18 +136,18 @@ export function WebappShortcutsRow({
               key={shortcut.url}
               shortcut={shortcut}
               appearance={appearance}
-              onEdit={onEdit}
-              onRemove={onRemove}
+              onEdit={manualRow.onEdit}
+              onRemove={manualRow.onRemove}
             />
           ))}
         </SortableContext>
       </DndContext>
-      {manager.canAdd && (
+      {manualRow.canAdd && (
         <AddShortcutTile
           appearance={appearance}
-          onClick={onAdd}
-          acceptsDroppedUrl={canAcceptDroppedUrl}
-          isDropActive={isDropTarget}
+          onClick={manualRow.onAdd}
+          acceptsDroppedUrl={manualRow.canAdd}
+          isDropActive={manualRow.isDropTarget}
         />
       )}
     </div>
