+++ packages/shared/src/features/shortcuts/hooks/useIsShortcutsHubEnabled.ts
@@ -1,0 +1,13 @@
+import { useAuthContext } from '../../../contexts/AuthContext';
+import { useConditionalFeature } from '../../../hooks/useConditionalFeature';
+import { featureShortcutsHub } from '../../../lib/featureManagement';
+
+export function useIsShortcutsHubEnabled(): boolean {
+  const { user } = useAuthContext();
+  const { value: hubEnabled } = useConditionalFeature({
+    feature: featureShortcutsHub,
+    shouldEvaluate: !!user,
+  });
+
+  return !!user && !!hubEnabled;
+}
